added the support of llava_v6 to the task

+++ lmms_eval/models/llava.py
@@ -1,5 +1,4 @@
 import torch
-
 torch.backends.cuda.matmul.allow_tf32 = True
 
 
@@ -96,11 +95,11 @@ class Llava(lmms):
         model_name = model_name if model_name is not None else get_model_name_from_path(pretrained)
         try:
             # Try to load the model with the multimodal argument
-            self._tokenizer, self._model, self._image_processor, self._max_length = load_pretrained_model(pretrained, None, model_name, device_map=self.device_map, **llava_model_args)
+            self._tokenizer, self._model, self._image_processor, self._max_length = load_pretrained_model(pretrained, None, model_name, device_map=self.device_map, torch_dtype="bfloat16", **llava_model_args)
         except TypeError:
             # for older versions of LLaVA that don't have multimodal argument
             llava_model_args.pop("multimodal", None)
-            self._tokenizer, self._model, self._image_processor, self._max_length = load_pretrained_model(pretrained, None, model_name, device_map=self.device_map, **llava_model_args)
+            self._tokenizer, self._model, self._image_processor, self._max_length = load_pretrained_model(pretrained, None, model_name, device_map=self.device_map, torch_dtype="bfloat16", **llava_model_args)
         self._config = self._model.config
         self.model.eval()
         if tie_weights:
@@ -225,9 +224,9 @@ class Llava(lmms):
             if visuals:
                 image = process_images(visuals, self._image_processor, self._config)
                 if type(image) is list:
-                    image = [_image.to(dtype=torch.float16, device=self.device) for _image in image]
+                    image = [_image.to(dtype=torch.bfloat16, device=self.device) for _image in image]
                 else:
-                    image = image.to(dtype=torch.float16, device=self.device)
+                    image = image.to(dtype=torch.bfloat16, device=self.device)
             else:
                 image = None
 
@@ -314,6 +313,7 @@ class Llava(lmms):
         chunks = re_ords.get_batched(n=self.batch_size, batch_fn=None)
         num_iters = len(requests) // self.batch_size if len(requests) % self.batch_size == 0 else len(requests) // self.batch_size + 1
         pbar = tqdm(total=num_iters, disable=(self.rank != 0), desc="Model Responding")
+        # breakpoint()
         for chunk in chunks:
             contexts, all_gen_kwargs, doc_to_visual, doc_id, task, split = zip(*chunk)
             task = task[0]
@@ -343,9 +343,9 @@ class Llava(lmms):
             if flattened_visuals:
                 image_tensor = process_images(flattened_visuals, self._image_processor, self._config)
                 if type(image_tensor) is list:
-                    image_tensor = [_image.to(dtype=torch.float16, device=self.device) for _image in image_tensor]
+                    image_tensor = [_image.to(dtype=torch.bfloat16, device=self.device) for _image in image_tensor]
                 else:
-                    image_tensor = image_tensor.to(dtype=torch.float16, device=self.device)
+                    image_tensor = image_tensor.to(dtype=torch.bfloat16, device=self.device)
             else:
                 image_tensor = None
 
@@ -432,7 +432,6 @@ class Llava(lmms):
             pbar.update(1)
             # reorder this group of results back to original unsorted form
         res = re_ords.get_original(res)
-
         pbar.close()
         return res
